pyproject.toml
[project]
name = "sunday"
version = "0.1.0"
description = "本地优先的个人边端 AI 智能体"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.4",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "pyyaml>=6.0",
    "textual>=0.80",
    "click>=8.1",
    "websockets>=13.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "vcrpy>=6.0",
    "ruff>=0.6",
]

[project.scripts]
sunday = "sunday.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sunday"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
